requirements.txt
matplotlib==3.6.2
numpy==1.23.5
opencv_contrib_python==4.8.1.78
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
pandas==2.0.0
Pillow==9.5.0
Pillow==9.0.1
Pillow==10.3.0
pycuda==2024.1
tensorrt==8.6.1.post1
torch==2.2.0+cu118
torchvision==0.17.0+cu118
tqdm==4.65.0
